Refactor merge & DML test helpers to allow reusability
